Add support for dynamic page cache directories
If a Rails application is multi-tenanted (i.e. serving multiple
domains from a single instance) then a global page cache directory
can be limiting. To help simplify this pattern this commit adds
support for using a Proc, Symbol or callable object that will
either be executed in the context of the controller, the method
called on the controller or the object called with the request
object respectively.

Fixes #36