Read content from response body when content is nil
The changes 43dd478 removed a `||` condition unintentionally than
used the response body if the content argument to `cache_page`
was nil. This commit reinstates that condition and expands the test
suite to assert on the content of the cached files rather than just
the existence of them which should prevent any similar regressions.

Fixes #42.